Move project-view into main directory since not shared
