+++ client/src/app/project-view/project-view.component.html
