+++ client/src/app/project-view/project-view.component.ts
@@ -1,10 +1,10 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {Project} from '../../models/project';
-import {AlertService} from '../../services/alert.service';
-import * as metadataSchema from '../../../project-form/project-metadata-schema.json';
-import * as ingestSchema from '../../../project-form/project-ingest-schema.json';
-import {layout} from '../../../project-form/layout';
-import {MetadataFormConfig} from '../../../metadata-schema-form/models/metadata-form-config';
+import {Project} from '../shared/models/project';
+import {AlertService} from '../shared/services/alert.service';
+import * as metadataSchema from '../project-form/project-metadata-schema.json';
+import * as ingestSchema from '../project-form/project-ingest-schema.json';
+import {layout} from '../project-form/layout';
+import {MetadataFormConfig} from '../metadata-schema-form/models/metadata-form-config';
 
 
 @Component({
