+++ client/src/app/project/project.component.html
@@ -17,8 +17,7 @@
   <br/>
   <app-upload *ngIf="upload" [projectUuid]="projectUuid"></app-upload>
   <br/>
-  <app-project-view (tabChange)="projectTabChange($event)" [project]="project">
-  </app-project-view>
+  <app-project-view (tabChange)="projectTabChange($event)" [project]="project"></app-project-view>
   <div class="dashboard" *ngIf="!upload && submissionEnvelopes && submissionEnvelopes.length > 0">
     <table>
       <thead class="thead-light">
